Use Integer for SuiteAuthorizationSource#getUser(id, ...) also. #1823.

+++ authorization/mock-plugin/src/main/java/gov/nih/nci/cabig/ctms/suite/authorization/plugins/mock/internal/StaticSuiteAuthorizationSource.java
@@ -38,7 +38,7 @@ public class StaticSuiteAuthorizationSource implements SuiteAuthorizationSource 
         return username.equals(soleUser.getUsername()) ? soleUser : null;
     }
 
-    public SuiteUser getUser(long id, SuiteUserRoleLevel desiredDetail) {
+    public SuiteUser getUser(int id, SuiteUserRoleLevel desiredDetail) {
         return id == soleUser.getId() ? soleUser : null;
     }
